+++ src/main/java/com/evacipated/cardcrawl/modthespire/lib/SpireOverride.java
@@ -1,0 +1,9 @@
+package com.evacipated.cardcrawl.modthespire.lib;
+
+import java.lang.annotation.*;
+
+@Target({ElementType.METHOD})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface SpireOverride
+{
+}