+++ src/main/java/com/evacipated/cardcrawl/modthespire/lib/SpireSuper.java
@@ -1,0 +1,11 @@
+package com.evacipated.cardcrawl.modthespire.lib;
+
+import sun.reflect.generics.reflectiveObjects.NotImplementedException;
+
+public class SpireSuper
+{
+    public static <R> R call(Object... params)
+    {
+        throw new NotImplementedException();
+    }
+}